+++ src/collective/documentgenerator/search_replace/pod_template.py
@@ -5,6 +5,8 @@ from plone.namedfile import NamedBlobFile
 import collections
 import mimetypes
 import os
+import shutil
+import tempfile
 
 
 SearchReplaceResult = collections.namedtuple(
@@ -28,11 +30,8 @@ class SearchAndReplacePODTemplates:
         """
         self.podtemplates = podtemplates
         self.templates_by_filename = {}
-        self.tmp_dir = "/tmp/docgen"
+        self.tmp_dir = tempfile.mkdtemp(prefix='docgen/')
         self.changed_files = set()
-
-        if not os.path.isdir(self.tmp_dir):
-            os.mkdir(self.tmp_dir)
 
         # compute the (future) file system path of the plone pod templates
         for podtemplate in podtemplates:
@@ -71,10 +70,8 @@ class SearchAndReplacePODTemplates:
                     filename=podtemplate.odt_file.filename,
                 )
                 podtemplate.odt_file = result
-        # clean tmp file
-        for filename in self.templates_by_filename.keys():
-            if os.path.isfile(filename):
-                os.remove(filename)
+        # delete tmp directory
+        shutil.rmtree(self.tmp_dir)
 
     def search(self, find_expr, is_regex=False):
         """