+++ CHANGES.rst
@@ -8,6 +8,10 @@ Changelog
   [sgeulette]
 - Added `get_relations` in dexterity
   [sgeulette]
+- Use Appy to search and replace in POD templates pod expressions.
+  [aduchene]
+- .ods POD Templates are now usable with search and replace.
+  [aduchene]
 
 3.29 (2022-04-15)
 -----------------